# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "API for managing referrals between business units",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "http://mytotalhealth.com.my/referral-api",
            "description": "MyHealth Server (Staging)"
        }
    ],
    "paths": {
...
        "/api/attachment/{attachment}": {
            "get": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Get attachment file as base64 encoded data",
                "description": "Get attachment file as base64",
                "operationId": "e3da93f887faf547d458b75b9d6154fb",
                "parameters": [
                    {
                        "name": "attachment",
                        "in": "path",
                        "description": "ID of the attachment to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "document.pdf"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "application/pdf"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 1024000
                                        },
                                        "base64": {
                                            "type": "string",
                                            "format": "byte",
                                            "example": "JVBERi0xLjQKJaqrrK0KMSAwIG9iag..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attachment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Attachment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}